requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram-tests==1.0.3
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.19.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
apache-libcloud==3.8.0
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
billiard==4.2.0
build==1.0.3
celery==5.3.6
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distlib==0.3.7
Django==5.0
django-celery-beat==2.6.0
django-phonenumber-field==7.2.0
docopt==0.6.2
environs==10.0.0
frozenlist==1.4.0
gevent==23.9.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.2
kombu==5.3.4
magic-filter==1.0.12
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.0
pep517==0.13.1
pep8-naming==0.13.3
phonenumbers==8.13.26
phonenumberslite==8.13.26
pika==1.3.2
pip-api==0.0.30
platformdirs==4.1.0
plette==0.4.4
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
pyproject_hooks==1.0.0
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requirementslib==3.0.0
ruff==0.1.8
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlalchemy-file==0.6.0
sqlparse==0.4.4
starlette==0.32.0.post1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvloop==0.19.0
vine==5.1.0
watchdog==3.0.0
watchdog-gevent==0.1.1
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
yarg==0.1.9
yarl==1.9.3
zope.event==5.0
zope.interface==6.1
pytest-env==1.1.3
django-configurations==2.5
gunicorn==21.2.0
django-import-export==4.0.0
